requirements.txt
streamlit>=1.34.0
spacy>=3.7.0
pdfplumber>=0.10.3
python-docx>=1.0.1
scikit-learn>=1.5.0
pandas>=2.2.0
joblib>=1.4.0
tqdm>=4.66.0
nltk>=3.8.1